refactor(access): tidy PanelAccess values and manager fixtures
PanelAccess never uses its string values, so plain auto() members
suffice. The manager persona fixtures move to conftest next to
staff_client for reuse.
