+++ tests/integration/conftest.py
@@ -231,6 +231,19 @@ def staff_client(client, staff_user):
     return client
 
 
+@pytest.fixture(name="manager_user")
+def manager_user_fixture(sphere):
+    user = UserFactory(username="manageruser", name="Manager User")
+    sphere.managers.add(user)
+    return user
+
+
+@pytest.fixture(name="manager_client")
+def manager_client_fixture(client, manager_user):
+    client.force_login(manager_user)
+    return client
+
+
 @pytest.fixture(params=["manager", "superuser"])
 def panel_access_user(request, active_user, sphere):
     if request.param == "manager":
